RPRBLND-1392: Update Athena for Autotests. (#34)
PURPOSE
Team from automation tests on CIS wants to check Athena data which is transferred to AWS. Therefore they asked to provide possibility not to remove corresponded athena json files after sending.

TECHNICAL STEPS
Added config setting clean_athena_files (True by default), by which we can control removing athena json files after sending statistics.

+++ src/rprblender/utils/athena.py
@@ -26,6 +26,7 @@ import bpy
 import pyrpr
 from rprblender import utils
 from rprblender import bl_info
+from rprblender import config
 
 from . import logging
 from . import IS_MAC
@@ -58,10 +59,11 @@ def _send_data_thread(data):
             is_error = True
 
         finally:
-            try:
-                os.remove(file_name)
-            except Exception as e:  # In case removal happens on Blender exit, when temporary folder is already removed
-                log.warn(f"Unable to remove temporary json: {e}")
+            if config.clean_athena_files:
+                try:
+                    os.remove(file_name)
+                except Exception as e:  # In case removal happens on Blender exit, when temporary folder is already removed
+                    log.warn(f"Unable to remove temporary json: {e}")
 
         log("send_data_thread finish")
